Rename hist to _hist
