+++ boost_histogram/_hist.py
